reordering config. removed workers restriction
Former-commit-id: 522ef0a10453b1a1e01a447e186fc17cd230ff4c

+++ include/IOWorker.h
@@ -41,7 +41,7 @@ namespace LAppS
       typedef WebSocket<TLSEnable,StatsEnable> WSType;
       typedef std::shared_ptr<WSType>          WSSPtr;
      
-    explicit IOWorker(const uint8_t id, const size_t maxConnections)
+    explicit IOWorker(const size_t id, const size_t maxConnections)
     : Worker(id,maxConnections), mMayRun(true),mCanStop(false),
       mConnectionsMutex(), mInboundMutex(), mOutMutex(),
       mShakespeer(),
